goose-guard-model/requirements.txt
pandas
transformers
transformers[torch]
torch
flask
python-dotenv